Rename formatter.py
